refactor to keep the same format for components
