+++ client/src/components/GameBar/GameBar.js
@@ -2,10 +2,8 @@ import React from "react";
 import { useHostId } from "../../contexts/DataContext";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
-import Avatar from "@material-ui/core/Avatar";
 import Box from "@material-ui/core/Box";
 import Button from "@material-ui/core/Button";
-import Icon from "@material-ui/core/Icon";
 import Typography from "@material-ui/core/Typography";
 import useStyles from "./styles";
 import ProfileBar from "../ProfileBar";
@@ -14,7 +12,7 @@ const GameBar = (props) => {
   const classes = useStyles(props);
   const [hostId] = useHostId();
   const userId = window.localStorage.getItem("id");
-  
+
   return (
     <nav className={classes.NavBar}>
       <Box className={classes.NavBarWrap}>
@@ -74,7 +72,7 @@ const GameBar = (props) => {
           >
             New Game
           </Button>
-          <ProfileBar history={props.history}/>
+          <ProfileBar history={props.history} />
         </Box>
       </Box>
     </nav>
